Moving the repositories into domain-specific packages for better organization
